Nuevo paquete: Role

+++ src/main/java/com/apms/humanResource/HumanResourceRepository.java
@@ -2,6 +2,6 @@ package com.apms.humanResource;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HumanResourceRepository extends JpaRepository<HumanResource, Long> {
+public interface HumanResourceRepository extends JpaRepository<HumanResource, Integer> {
 
 }
