+++ src/main/java/com/apms/title/TitleRepository.java
@@ -2,6 +2,6 @@ package com.apms.title;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TitleRepository extends JpaRepository<Title, Long> {
+public interface TitleRepository extends JpaRepository<Title, Integer> {
 
 }
