+++ src/main/java/com/apms/workplace/WorkplaceRepository.java
@@ -2,6 +2,6 @@ package com.apms.workplace;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkplaceRepository extends JpaRepository<Workplace, Long> {
+public interface WorkplaceRepository extends JpaRepository<Workplace, Integer> {
 
 }
