+++ src/main/java/com/apms/workplaceType/WorkplaceTypeRepository.java
@@ -2,6 +2,6 @@ package com.apms.workplaceType;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkplaceTypeRepository extends JpaRepository<WorkplaceType, Long> {
+public interface WorkplaceTypeRepository extends JpaRepository<WorkplaceType, Integer> {
 
 }
